refactor, remove duplicated code
