+++ pkg/daemon/daemon.go
@@ -3,15 +3,35 @@ package daemon
 import (
 	"context"
 	"fmt"
+	"log"
 	"os"
 	"sync"
+	"time"
 
 	"github.com/queryplan-ai/queryplan-proxy/pkg/daemon/types"
+	"github.com/queryplan-ai/queryplan-proxy/pkg/heartbeat"
 	"github.com/queryplan-ai/queryplan-proxy/pkg/mysql"
 	"github.com/queryplan-ai/queryplan-proxy/pkg/postgres"
 )
 
+const (
+	sendInterval = 10 * time.Second
+)
+
 func Run(ctx context.Context, opts types.DaemonOpts) {
+	go func() {
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			case <-time.After(sendInterval):
+				if err := heartbeat.SendPendingQueries(ctx, opts); err != nil {
+					log.Printf("Error sending pending queries: %v", err)
+				}
+			}
+		}
+	}()
+
 	switch opts.DBMS {
 	case types.Postgres:
 		var wg sync.WaitGroup
