+++ pkg/heartbeat/heartbeat.go
@@ -1,0 +1,16 @@
+package heartbeat
+
+import (
+	"time"
+
+	"github.com/queryplan-ai/queryplan-proxy/pkg/heartbeat/types"
+)
+
+func AddPendingQuery(query string) {
+	qpq := types.QueryPlanQuery{
+		Query:      query,
+		ExecutedAt: time.Now().UnixNano(),
+	}
+
+	pendingQueries.Add(qpq)
+}
