+++ pkg/postgres/proxy.go
@@ -6,28 +6,9 @@ import (
 	"io"
 	"log"
 	"net"
-	"sync"
-	"time"
 
 	daemontypes "github.com/queryplan-ai/queryplan-proxy/pkg/daemon/types"
-	"github.com/queryplan-ai/queryplan-proxy/pkg/mysql/types"
-	"github.com/queryplan-ai/queryplan-proxy/pkg/ringbuffer"
-)
-
-const (
-	sendInterval = 10 * time.Second
-)
-
-var (
-	queryRegistry sync.Map
-)
-
-const (
-	defaultMaxPendingQueriesSize = 10000
-)
-
-var (
-	pendingQueries = ringbuffer.New[types.QueryPlanQuery](defaultMaxPendingQueriesSize)
+	"github.com/queryplan-ai/queryplan-proxy/pkg/heartbeat"
 )
 
 func RunProxy(ctx context.Context, opts daemontypes.DaemonOpts) {
@@ -42,19 +23,6 @@ func RunProxy(ctx context.Context, opts daemontypes.DaemonOpts) {
 	upstreamAddress := fmt.Sprintf("%s:%d", opts.UpstreamAddress, opts.UpstreamPort)
 
 	fmt.Printf("Listening on %s, proxying to %s\n", address, upstreamAddress)
-
-	go func() {
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case <-time.After(sendInterval):
-				if err := sendPendingQueries(ctx, opts); err != nil {
-					log.Printf("Error sending pending queries: %v", err)
-				}
-			}
-		}
-	}()
 
 	for {
 		select {
@@ -119,11 +87,7 @@ func (w *loggingWriter) Write(p []byte) (n int, err error) {
 			if err != nil {
 				log.Printf("Error cleaning query: %v", err)
 			} else {
-				qpq := types.QueryPlanQuery{
-					Query:      cleanedQuery,
-					ExecutedAt: time.Now().UnixNano(),
-				}
-				pendingQueries.Add(qpq)
+				heartbeat.AddPendingQuery(cleanedQuery)
 			}
 		}
 	}
